services/webhook-listener/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
aioredis
aiohttp
